hrms: int to Long search criteria

+++ rainmaker/egov-hrms/src/main/java/org/egov/hrms/web/contract/EmployeeSearchCriteria.java
@@ -39,7 +39,7 @@ public class EmployeeSearchCriteria {
 	
 	public List<String> uuids;
 	
-	public List<Integer> positions;
+	public List<Long> positions;
 	
 	public Boolean isActive;
